requirements.txt
asgiref==3.2.7
Babel==2.7.0
Django==2.2.10
django-bootstrap-modal-forms==1.5.0
django-phonenumber-field==3.0.1
django-pwa==1.0.5
lxml==4.3.4
phonenumbers==8.12.1
phonenumberslite==8.10.14
Pillow==6.0.0
python-docx==0.8.10
pytz==2019.1
sqlparse==0.3.0
